refactor: 핸들러 로직 이동
