+++ src/communicator/context/connection.tsx
@@ -1,7 +1,12 @@
 import { createContext, FC, ReactNode, useContext } from "react";
+import { useSetRecoilState } from "recoil";
+
+import { connectionStateAtom } from "@/states/connection";
+import { joinStatusAtom } from "@/states/join";
 
 import { Connection } from "../types";
 import { notInitializedObject } from "../util";
+import { useSendMessage } from "./communicator";
 
 interface ConnectionContext {
   connection: Connection;
@@ -34,12 +39,28 @@ export const ConnectionProvider: FC<ConnectionProviderProps> = ({
 export const useConnection = () => {
   const { connection } = useContext(ConnectionContext);
 
+  const send = useSendMessage();
+  const setJoinStatus = useSetRecoilState(joinStatusAtom);
+  const setConnectionStatus = useSetRecoilState(connectionStateAtom);
+
   return {
     async connect() {
       await connection.connect();
     },
     async disconnect() {
       await connection.disconnect();
+    },
+    async join(nickname: string) {
+      setJoinStatus({
+        state: "loading",
+      });
+      setConnectionStatus({ status: "connecting" });
+      await connection.connect();
+      setConnectionStatus({ status: "connected" });
+
+      send("join", {
+        nickname,
+      });
     },
     setOnDisconnect(callback: () => void) {
       connection.onDisconnect(callback);
